Begin laying foundation for moving tabs around + instanced tabs
