+++ src/components/contexts/PatientListContext.jsx
@@ -1,7 +1,4 @@
-import React, { createContext, useState, useContext } from 'react';
+import React, { createContext, useContext } from 'react';
 
-// Create a context for patient lists data
 export const PatientListsContext = createContext();
-
-// Custom hook to use the patient lists context
 export const usePatientLists = () => useContext(PatientListsContext);
